Use a low priority otlp protocol configuration customizer
