+++ javaagent-tooling/src/test/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.AutoConfigurationCustomizerProvider
@@ -1,2 +1,3 @@
 io.opentelemetry.javaagent.tooling.config.ConfigurationPropertiesSupplier
+io.opentelemetry.javaagent.tooling.config.OtlpProtocolPropertiesSupplier
 io.opentelemetry.javaagent.tooling.config.ConfigurationPropertiesSupplierTest$UserCustomPropertiesSupplier
